backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
environs==14.5.0
filelock==3.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
marshmallow==4.2.2
packaging==25.0
pipenv==2026.0.3
platformdirs==4.5.1
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==3.0
PyGithub==2.9.0
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
requests==2.33.1
setuptools==80.9.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
